+++ server/package.json
@@ -4,8 +4,8 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "dev": "nodemon app",
-    "start": "NODE_ENV=production pm2 start app.js"
+    "dev": "nodemon src/index",
+    "start": "NODE_ENV=production pm2 start src/index.js"
   },
   "author": "By-juun",
   "license": "MIT",
