+++ server/src/index.js
@@ -1,0 +1,7 @@
+const express = require("../config/express");
+
+function bootStrap() {
+  express();
+}
+
+bootStrap();
